# Conflicts:
#	Assets/Scenes/combat-system.unity
#	Assets/Scripts/General/IDamageable.cs
#	Assets/Scripts/architecture.cs

+++ Assets/Scripts/General/IDamageable.cs
@@ -3,5 +3,5 @@
     // Called by Player (for AIEnemy) and by AIEnemy (for Building)
     bool IsDead();
     // Called by Player (for AIEnemy) and by AIEnemy (for Building)
-    void TakeDamage(float damage, AttackType attacktype);
+    void TakeDamage(int damage, AttackType attacktype);
 }
